+++ CHANGELOG.md
@@ -9,5 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Logs service.
 - Log all proxied requests to the file 'logs/http.log'.
 - New scripts in package.json for start, stop and update.
+- Store pid in .pid file.
+- Npm script for stop process using the pid stored in .pid file.
+- SGTERM and SIGINT signal handlers.
 
 ### Fixed